taskUp/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
alembic==1.12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.3.0
email-validator==2.0.0
websockets==11.0.3
pytest==7.4.2
httpx==0.24.1